pyteal: Add stateful smart contract implementation.
This is the PyTEAL implementation of a stateful smart contract that
allows to deposit any custom token using a single deposit transaction
with rekeyTo field set by the smart contract caller to the application.

The smart contract source code is well documented. The demonstrantion and
usage will be provided in the scripts folder of the example.